+++ src/trabalho1/Pessoa.java
@@ -3,9 +3,9 @@ package trabalho1;
 
 public abstract class Pessoa {
     public String nome;
-    public int cpf;
+    public long cpf;
 
-    public Pessoa(String nome, int cpf) {
+    public Pessoa(String nome, long cpf) {
         this.nome = nome;
         this.cpf = cpf;
     }
